pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-db"
version = "1.0.0"
dependencies = [
    "openai",
    "pydantic",
    "python-dotenv",
    "faker",
    "fastapi",
    "uvicorn",
    "httpx",
    "openenv-core"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]